remediation/webhook-receiver/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pyyaml==6.0.2
